Refactoring - moving netconf related dependencies from core implementation to cisco-xr-driver
