+++ features/odl-unimgr/pom.xml
@@ -30,7 +30,7 @@
     <netvirt.version>1.5.0</netvirt.version>
     <vpnservice.version>0.5.0</vpnservice.version>
     <mdsal.model.version>0.11.0</mdsal.model.version>
-    <netconf.version>1.3.0</netconf.version>
+
     <genius.version>0.3.0</genius.version>
     <l2switch.version>0.6.0</l2switch.version>
     <configfile.directory>etc/opendaylight/karaf</configfile.directory>
@@ -72,27 +72,6 @@
       <groupId>org.opendaylight.unimgr</groupId>
       <artifactId>odl-unimgr-api</artifactId>
       <version>${project.version}</version>
-      <type>xml</type>
-      <classifier>features</classifier>
-    </dependency>
-    <dependency>
-      <groupId>org.opendaylight.netconf</groupId>
-      <artifactId>odl-netconf-connector</artifactId>
-      <version>${netconf.version}</version>
-      <type>xml</type>
-      <classifier>features</classifier>
-    </dependency>
-    <dependency>
-      <groupId>org.opendaylight.netconf</groupId>
-      <artifactId>odl-netconf-connector-ssh</artifactId>
-      <version>${netconf.version}</version>
-      <type>xml</type>
-      <classifier>features</classifier>
-    </dependency>
-    <dependency>
-      <groupId>org.opendaylight.netconf</groupId>
-      <artifactId>odl-netconf-tcp</artifactId>
-      <version>${netconf.version}</version>
       <type>xml</type>
       <classifier>features</classifier>
     </dependency>
