--- impl/src/main/java/org/opendaylight/unimgr/utils/DriverConstants.java
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) 2017 Cisco Systems Inc and others.  All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v1.0 which accompanies this distribution,
- * and is available at http://www.eclipse.org/legal/epl-v10.html
- */
-
-package org.opendaylight.unimgr.utils;
-
-/**
- * @author bartosz.michalik@amartus.com
- */
-public interface DriverConstants {
-    String XR_NODE = "xr-node";
-}
